Subzero: Render constants in dump() to be more like LLVM.
Integers are generally dumped as signed instead of unsigned values.
Integers of i1 type are dumped as 'false' and 'true'.  Floating point
values still don't match LLVM.

BUG= none
R=kschimpf@google.com

Review URL: https://codereview.chromium.org/539743002

+++ src/IceInst.cpp
@@ -663,8 +663,8 @@ void InstSwitch::dump(const Cfg *Func) const {
   getSrc(0)->dump(Func);
   Str << ", label %" << getLabelDefault()->getName() << " [\n";
   for (SizeT I = 0; I < getNumCases(); ++I) {
-    Str << "    " << Ty << " " << getValue(I) << ", label %"
-        << getLabel(I)->getName() << "\n";
+    Str << "    " << Ty << " " << static_cast<int64_t>(getValue(I))
+        << ", label %" << getLabel(I)->getName() << "\n";
   }
   Str << "  ]";
 }
